Rename project files from ConsumerSurveysSample to GoogleSurveysSample.
