--- client_libraries-DEPRECATED/csharp/src/ConsumerSurveysSample/packages.config
@@ -1,17 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="BouncyCastle" version="1.8.1" targetFramework="net452" />
-  <package id="CommandLineParser" version="2.0.233-beta" targetFramework="net452" />
-  <package id="Google.Apis" version="1.19.0" targetFramework="net452" />
-  <package id="Google.Apis.Auth" version="1.19.0" targetFramework="net452" />
-  <package id="Google.Apis.Core" version="1.19.0" targetFramework="net452" />
-  <package id="Google.Apis.Surveys.v2" version="1.19.0.672" targetFramework="net452" />
-  <package id="log4net" version="2.0.5" targetFramework="net452" />
-  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net452" />
-  <package id="Microsoft.Bcl.Async" version="1.0.168" targetFramework="net452" />
-  <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="net452" />
-  <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="net452" />
-  <package id="NDesk.Options" version="0.2.1" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net452" />
-  <package id="Zlib.Portable.Signed" version="1.11.0" targetFramework="net452" />
-</packages>
